Resolve many Win32-specific compilation issues
* No ssize_t on windows (supports standard ptrdiff_t)
* Assumptions that int maps to 64 bits (as on linux)
* Windows #define'd conflicts with common keywords
* #preprocessor evaluation is not valid within preprocessor macro args


Roll-up WIP of the Windows Port of Envoy

See below for the many authors of this fork. Notable changes include;

- libevent does not support EV_ET on windows, so we make it an error to
use FileTriggerType::Edge on Windows

- runfiles for tests are not symlinked, so we need to look up their
location in a MANIFEST file

- On Windows, when two listeners use the same cert and session ticket key, but
different client CA, the session cannot be resumed between the two. This is the correct
behavior. However, the exact error the server gets is slightly different
(SSL_ERROR_SYSCALL on Windows vs. SSL_ERROR_SSL with TLSV1_ALERT_UNKNOWN_CA on Linux)
The tests are updated to reflect this

- we needed to add separate methods for writeFile / writeSocket and
closeFile / closeSocket to handle the fact that files and sockets have
different types on windows (int vs. SOCKET)

Pass EVLOOP_NONBLOCK | EVLOOP_ONCE to event_base_loop.
This results in the expected behavior for NonBlock

Ensure the capture transport socket writes binary files in binary mode

Route all syscalls through Api::OsSysCallsImpl

Split os_sys_calls implementation into linux and win32 packages

Add implementation for Win32 filewatcher

This uses the ReadDirectoryChangesW API. Since this API does not work at
all with libevent, use a socket on localhost and a conccurrent queue to
synchronize the execution of the callbacks

Change nghttp2 static lib name to nghttp2_static
This is because their build system also produces an nghttp2.lib file
that dynamically links to nghttp2.dll; we do not want to use that one

Split thread impl into separate win32 and pthread

Use sizeof instead of hardcoding count

Fix 3 integration tests
They were failing due to trying to connect to the any address,
use loopback instead

Make sure MSVC defines __cplusplus

Fix ASSERT macro with multiple args when using MSVC

Replace strtol and strtoul with strtoll and strtoull

long / unsigned long are only guaranteed to be at least 32-bits, and on Windows
they are just that. Use long long / unsigned long long to guarantee 64-bit range

Don't append buffer if strftime returns 0

There seems to be an issue with absl::StrCat in this case, so we just
work around it

Fix regex logic in access log formatter

Add default move constructor to SymbolEncoding
This allows a returned SymbolEncoding object to be moved instead
of copied, which prevents an assertion failure when the copied object is
destructed

Patch nghttp2 so that ssize_t is correct on Windows

Correct linkage for nghttp2_enable_strict_preface

Replaced all `SOCKET_HANDLE`s with `Network::IoHandle`s

define missing IPV6_RECVPKTINFO macro in windows

patch tclap to account for win64 size_t increase to long long

Correct emplace_back to convert to an unsigned int system status code (32 bits)

Toggle a number of cmake-based dependencies to compile /MT against libcmt.lib (static)

Resolve missing sendto() function on win32

Remove support for wide characters from the chromium url handler
This existed in the original code but is never used by envoy. This code is all to be replaced
in the future according to;
envoyproxy@c22cfd2

Windows uses Sleep (no usleep etc.)

Comment out problem gcc flags for msvc compilation

Small change to envoy_cc_build() to carry tags[] along

Also take tags for the protoc rules to allow us to skip_on_windows

Work around missing posix PATH_MAX on Windows

Drop tests with pipe implementation problems for now
TODO: PipeInstance (not implemented on Windows yet)

Manipulate var_name as a std::string for Windows envvar handling

Bump PGV to Sep 4th
Includes envoyproxy/protoc-gen-validate#272

ptrdiff_t is a C99 stdint.h type - substitute ssize correctly
The posix ssize_t has never been standard C but we need to consume nghttp2 and libevent
which make heavy use of these sizes in public headers. Use this type on windows over
other Windows-specific types such as SSIZE_T.

+++ test/server/options_impl_test.cc
@@ -14,6 +14,7 @@
 
 #if defined(__linux__)
 #include <sched.h>
+#include "common/api/os_sys_calls_impl_linux.h"
 #include "server/options_impl_platform_linux.h"
 #endif
 #include "test/mocks/api/mocks.h"
@@ -40,7 +41,7 @@ public:
     std::transform(words.cbegin(), words.cend(), std::back_inserter(argv),
                    [](const std::string& arg) { return arg.c_str(); });
     return std::make_unique<OptionsImpl>(
-        argv.size(), argv.data(), [](bool) { return "1"; }, spdlog::level::warn);
+        static_cast<int>(argv.size()), argv.data(), [](bool) { return "1"; }, spdlog::level::warn);
   }
 };
